requirements.txt
img2pdf==0.5.1
pillow==11.0.0
PyMuPDF==1.24.10
PyPDF2==3.0.1
